ZCS-10322:Support for Attachment Type Restrictions
Added ldap attribute zimbraFeatureFileTypeUploadRestrictionsEnabled
for enabling and disabling the blocked file types set in the
zimbraFileUploadBlockedFileTypes for uploading. If this ldap attribute
is disabled then the atachment is uploaded without scanning the ldap
attribute zimbraFileUploadBlockedFileTypes and the any type of the
attachments are allowed to be uploaded.

If the ldap attribute zimbraFeatureFileTypeUploadRestrictionsEnabled
is enabled then the file extensions mentioned in the ldap attribute
are scanned one by one and if any of the matching blocked extensions
are found during attachement upload then BLOCKED_FILE_TYPE_UPLOAD
service exception is thrown and the attachment is disallowed from
uploading. And suppose in case while uploading the attachement the
extension is not found in the blocked zimbraFileUploadBlockedFileTypes
then after that it's content-type is detected using the tika-core
library and based on it's content type the file extension is detected
and matched from the already fetched list of the blocked extension
and if a match is found then BLOCKED_FILE_TYPE_UPLOAD service exception
is thrown and the attachment is disallowed from uploading.